+++ mongoengine/base/fields.py
@@ -134,7 +134,9 @@ class BaseField:
         # If setting to None and there is a default value provided for this
         # field, then set the value to the default value.
         if value is None:
-            if not self.null and self.default is not None:
+            if self.null:
+                value = None
+            elif self.default is not None:
                 value = self.default
                 if callable(value):
                     value = value()
@@ -281,16 +283,15 @@ class ComplexBaseField(BaseField):
         return documents
 
     def __set__(self, instance, value):
-        unprocessable_fields = (
-            ComplexBaseField,
-            _import_class("EmbeddedDocumentField"),
-            _import_class("FileField"),
-        )
-        if self.field and not isinstance(self.field, unprocessable_fields):
+        # Some fields e.g EnumField are converted upon __set__
+        # So it is fair to mimic the same behavior when using e.g ListField(EnumField)
+        EnumField = _import_class("EnumField")
+        if self.field and isinstance(self.field, EnumField):
             if isinstance(value, (list, tuple)):
                 value = [self.field.to_python(sub_val) for sub_val in value]
             elif isinstance(value, dict):
                 value = {key: self.field.to_python(sub) for key, sub in value.items()}
+
         return super().__set__(instance, value)
 
     def __get__(self, instance, owner):
@@ -445,7 +446,7 @@ class ComplexBaseField(BaseField):
                             " have been saved to the database"
                         )
 
-                    # If its a document that is not inheritable it won't have
+                    # If it's a document that is not inheritable it won't have
                     # any _cls data so make it a generic reference allows
                     # us to dereference
                     meta = getattr(v, "_meta", {})